Refactor sweep script into a modular package
Break up the 1,749-line monolithic sweep.py into a well-organized
package with focused modules:

- constants.py: Exceptions, default search spaces, net-arch presets
- search_space.py: Loading, resolution, and per-stage extraction
- trial.py: Single trial execution and parameter spec building
- results.py: Trial result collection, CSV export, plotting, best-trial selection
- state.py: Sweep state persistence (local + GCS save/load/resume)
- submit.py: Vertex AI HPT job submission with retry logic
- orchestration.py: Single-stage and multi-stage launch coordination
- __main__.py: CLI argument parser and entry point
- __init__.py: Public API re-exports for backward compatibility

Each module stays under ~300 lines. All 67 existing tests pass
unchanged since __init__.py re-exports every symbol.

Docker/Vertex AI trial command updated from direct file execution
to `python -m environments.shared.scripts.sweep`.
